+++ SmashDBCore/pom.xml
@@ -5,7 +5,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>com.nk</groupId>
+    <groupId>nk.ssb</groupId>
     <artifactId>SmashDB</artifactId>
     <version>1.0-SNAPSHOT</version>
   </parent>
