+++ SmashDBCore/src/main/java/nk/ssb/smashdb/core/SmashDBStyle.java
@@ -1,4 +1,4 @@
-package nk.smashdb.core;
+package nk.ssb.smashdb.core;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
