+++ SmashDBCore/src/main/java/nk/ssb/smashdb/core/UserEggIF.java
@@ -1,4 +1,4 @@
-package nk.smashdb.core;
+package nk.ssb.smashdb.core;
 
 import org.immutables.value.Value;
