+++ SmashDBService/src/main/java/nk/ssb/smashdb/service/SmashDBConfiguration.java
@@ -1,4 +1,4 @@
-package nk.smashdb.service;
+package nk.ssb.smashdb.service;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
